Move role updates to an Admin Action
