+++ public/app/themes/clarity/inc/admin/users/add-agency-editor.php
@@ -4,6 +4,10 @@
  *
  * Agency Editor:
  * Third greatest permission group.
+ * 
+ * If you edit this file, you must sync the roles to the database by running the 
+ * SyncUserRoles admin action. 
+ * Navigate to Tools > Admin Commands > Sync user roles from codebase to database
  *
  * @package Clarity
  */
